+++ webflows/src/main/kotlin/com/schibsted/account/webflows/token/IdTokenValidator.kt
@@ -74,7 +74,7 @@ internal object IdTokenValidator {
                 claims.subject,
                 claims.getStringClaim("legacy_user_id"),
                 claims.audience,
-                (claims.expirationTime.time / 1000).toInt(),
+                (claims.expirationTime.time / 1000),
                 claims.getStringClaim("nonce"),
                 claims.getStringListClaim("amr")
             )
